+++ frontend/fingertips-frontend/components/organisms/LineChartTable/index.tsx
